move files from quantization/prototype -> prototype/quantization
