+++ torchao/prototype/quantization/mixed_precision/scripts/naive_intNwo.py
