+++ torchao/prototype/quantization/qat/affine_fake_quantized_tensor.py
