+++ torchao/prototype/quantization/qat/images/qat_diagram.png
